converters are no longer allowed on const&
cast can't return a reference to a temporary that was created within the
cast function. Since the conversion creates a brand new value within the
cast function, returning a ref to that is a no-go. We also don't know
whether a conversion is going to happen until runtime which means that
we can't change the function signature to suit our needs.

Long story short, we're fucked.